wip: extract type
