+++ src/routing/route/route-declaration.type.ts
@@ -1,0 +1,13 @@
+import type { HttpMiddleware, HttpRequest } from '@/Http';
+import type { RouteOptions } from '@/routing/route/route-options.type';
+import type { HttpMethod } from '@/routing/verb/http-method.type';
+import type { Request } from 'koa';
+
+export interface RouteDeclaration<TRequest extends Request = HttpRequest> {
+  name?: string;
+  path: string;
+  method: HttpMethod | HttpMethod[];
+  handler: HttpMiddleware<TRequest>;
+  middleware?: HttpMiddleware<TRequest>[];
+  options?: RouteOptions;
+}
